Move tests out of __tests__ directory
