+++ plugins/rich-editor/src/scripts/quill/KeyboardBindings.test.ts
@@ -5,8 +5,8 @@
  */
 
 import Delta from "quill-delta";
-import Quill from "../../quill";
-import KeyboardBindings from "../KeyboardBindings";
+import Quill from "../quill";
+import KeyboardBindings from "./KeyboardBindings";
 import { expect } from "chai";
 const LINE_FORMATS = ["blockquote-line", "spoiler-line"];
