+++ plugins/rich-editor/src/scripts/quill/rendering.test.ts
